transferred all components into components folder
