--- client/src/components/pages/Cart.js
@@ -1,7 +1,0 @@
-export const Cart = () => {
-	return(
-		<>
-			<h1>This is the cart page</h1>
-		</>
-	);
-}
